requirements.txt
arabic-reshaper==3.0.0
asn1crypto==1.5.1
aspose-cells==25.7.0
Aspose.Slides==24.6.0
babel==2.17.0
blinker==1.9.0
Brotli==1.1.0
certifi==2025.7.14
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.2
click==8.2.1
cobble==0.1.4
colorama==0.4.6
cryptography==45.0.5
cssselect2==0.8.0
Deprecated==1.2.18
distro==1.9.0
et_xmlfile==2.0.0
fire==0.7.0
Flask==2.3.3
Flask-Cors==4.0.0
fonttools==4.59.0
gunicorn==21.2.0
html5lib==1.1
idna==3.10
img2pdf==0.5.1
itsdangerous==2.2.0
Jinja2==3.1.6
jpype1==1.6.0
lxml==6.0.0
mammoth==1.6.0
MarkupSafe==3.0.2
numpy==2.2.6
opencv-python-headless==4.12.0.88
openpyxl==3.1.2
oscrypto==1.3.0
packaging==25.0
pandas==2.2.2
pdf2docx==0.5.8
pdf2image==1.17.0
pdfminer.six==20221105
pdfplumber==0.10.2
pikepdf==8.10.0
pillow==10.2.0
pycparser==2.22
pydyf==0.11.0
pyHanko==0.29.1
pyhanko-certvalidator==0.27.0
PyMuPDF==1.23.24
pypdf==5.9.0
PyPDF2==3.0.1
pypdfium2==4.30.0
pyphen==0.17.2
python-bidi==0.6.6
python-dateutil==2.9.0.post0
python-docx==1.1.0
python-magic==0.4.27
python-pptx==0.6.21
pytz==2025.2
PyYAML==6.0.2
reportlab==4.0.9
requests==2.32.4
six==1.17.0
svglib==1.5.1
termcolor==3.1.0
tinycss2==1.4.0
tqdm==4.67.1
typing_extensions==4.14.1
tzdata==2025.2
tzlocal==5.3.1
uritools==5.0.0
urllib3==2.5.0
weasyprint==60.2
webencodings==0.5.1
Werkzeug==2.3.7
wrapt==1.17.2
xhtml2pdf==0.2.17
xlsxwriter==3.2.5
zopfli==0.2.3.post1
easyocr
torch
rembg
onnxruntime